requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.90.1
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
pydantic==1.10.4
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.23.1
typing-extensions==4.4.0
uuid==1.30
uvicorn==0.20.0